DiT/requirements.txt
absl-py==2.2.2
accelerate==1.4.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anykeystore==0.2
apex==0.1
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==21.4b2
boto3==1.39.2
botocore==1.39.2
braceexpand==0.1.7
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colossalai==0.3.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contexttimer==0.3.3
contourpy==1.3.0
cryptacular==1.6.2
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.2
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1744321239219/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.32.2
distlib==0.3.9
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
fabric==3.2.2
facexlib==0.3.0
fairscale==0.4.13
faiss-gpu==1.7.2
filelock==3.17.0
filterpy==1.4.5
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.71.0
h5py==3.13.0
hf-xet==1.1.7
html5lib==1.1
huggingface-hub==0.34.4
hupper==1.12.1
hydra-core==1.3.2
identify==2.6.12
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources==6.5.2
invoke==2.2.0
iopath==0.1.9
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1701831663892/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
keras==2.15.0
kiwisolver==1.4.7
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.14.3
llvmlite==0.43.0
lmdb==1.6.2
looseversion==1.3.0
lpips==0.1.4
lvis==0.5.3
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mda-xdrlib==0.2.0
mdurl==0.1.2
ml-dtypes==0.2.0
mmcv==1.7.0
mmengine==0.10.7
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.2.1
nibabel==5.3.2
ninja==1.11.1.4
nodeenv==1.9.1
numba==0.60.0
numexpr==2.10.2
numpy==1.24.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
PasteDeploy==3.1.0
pathspec==0.12.1
pbkdf2==1.3
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1742485085/work
portalocker==3.2.0
pre_commit==4.2.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
propcache==0.3.1
protobuf==4.25.6
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1740663125313/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.8
pycortex==1.2.11
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydot==4.0.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyiqa==0.1.10
PyNaCl==1.5.0
pyparsing==3.2.1
pyramid==2.0.2
pyramid-mailer==0.15.1
PySnooper==1.2.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python3-openid==3.2.0
pytorch-fid==0.3.0
pytz==2025.1
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1743831216818/work
regex==2024.11.6
repoze.sendmail==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rotary-embedding-torch==0.8.7
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.25.1
setproctitle==1.3.5
shapely==2.0.7
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
soupsieve==2.7
SQLAlchemy==2.0.41
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.2.1
torch==2.6.0+cu118
torch-geometric==2.6.1
torchmetrics==1.7.3
TorchSnooper==0.8
torchvision==0.21.0+cu118
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615921868/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transaction==5.0
transformers==4.33.2
translationstring==1.4
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
vbench==0.1.5
velruse==1.1.1
venusian==3.1.1
virtualenv==20.31.2
wandb==0.19.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webdataset==0.2.100
webencodings==0.5.1
WebOb==1.8.9
Werkzeug==3.1.3
wrapt==1.14.1
WTForms==3.2.1
wtforms-recaptcha==0.3.2
xformers==0.0.29.post2
yacs==0.1.8
yapf==0.43.0
yarl==1.19.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zope.deprecation==5.1
zope.interface==7.2
zope.sqlalchemy==3.1
